#53 Move CORS settings for slack archive to CorsConfig and add it to dev profile
