+++ slack-archive/src/main/resources/application.properties
@@ -1,3 +1,4 @@
+spring.profiles.active=dev
 spring.output.ansi.enabled=ALWAYS
 #Mongo DB
 spring.data.mongodb.uri=mongodb://root:root@localhost:27017/slack-archive
